Improve logic in Command class
Now arguments are only appended to command string when
ExecuteCommandAsync is called, and features some type checking
against the passed arguments.
Argument info is stored in a specific struct and can be individually
queried from the Command.
ExecuteCommandAsync now returns struct with info of command execution,
including the exit code, total execution time and output of the process.